requirements.txt
ffmpeg-python
numpy
torch
torchaudio
librosa
scikit-learn
pandas
matplotlib
speechbrain